Run Rust actions in pre-built containers
Instead of running Rust actions in the default Ubuntu runner, which
requires us to install Rust manually, we are instead choosing to run the
actions in a pre-built Docker container. The container is built from
the official Rust container[^1], but contains some additional tools used
in tests.

[^1]: https://hub.docker.com/_/rust

Build Docker image for Rust actions

Use custom Docker container for Rust actions